requirements.txt
opencv-python-headless
easyocr
streamlit
Pillow
